+++ ibllib/io/spikeglx.py
@@ -94,6 +94,14 @@ class Reader:
             return self.read(nsel=item[0], csel=item[1], sync=False)
 
     @property
+    def geometry(self):
+        """
+        Gets the geometry, ie. the full trace header for the recording
+        :return: dictionary with keys 'row', 'col', 'ind', 'shank', 'adc', 'x', 'y', 'sample_shift'
+        """
+        return _geometry_from_meta(self.meta)
+
+    @property
     def shape(self):
         return self.ns, self.nc
 
@@ -107,10 +115,13 @@ class Reader:
 
     @property
     def version(self):
-        """:return: """
-        if not self.meta:
-            return None
-        return _get_neuropixel_version_from_meta(self.meta)
+        """Gets the version string: '3A', '3B2', '3B1', 'NP2.1', 'NP2.4'"""
+        return None if self.meta is None else _get_neuropixel_version_from_meta(self.meta)
+
+    @property
+    def major_version(self):
+        """Gets the the major version int: 1 or 2"""
+        return None if self.meta is None else _get_neuropixel_major_version_from_meta(self.meta)
 
     @property
     def rl(self):
@@ -378,13 +389,19 @@ def _get_serial_number_from_meta(md):
         return int(serial)
 
 
+def _get_neuropixel_major_version_from_meta(md):
+    MAJOR_VERSION = {'3A': 1, '3B2': 1, '3B1': 1, 'NP2.1': 2, 'NP2.4': 2.4}
+    version = _get_neuropixel_version_from_meta(md)
+    if version is not None:
+        return MAJOR_VERSION[version]
+
+
 def _get_neuropixel_version_from_meta(md):
     """
     Get neuropixel version tag (3A, 3B1, 3B2) from the metadata dictionary
     """
     if 'typeEnabled' in md.keys():
         return '3A'
-
     prb_type = md.get('imDatPrb_type')
     # Neuropixel 1.0 either 3B1 or 3B2 (ask Olivier about 3B1)
     if prb_type == 0:
@@ -451,6 +468,24 @@ def _get_type_from_meta(md):
         return 'ap'
     elif snsApLfSy == [-1, -1, -1] and md.get('typeThis', None) == 'nidq':
         return 'nidq'
+
+
+def _geometry_from_meta(meta_data):
+    """
+    Gets the geometry, ie. the full trace header for the recording
+    :param meta_data: meta_data dictionary as read by ibllib.io.spikeglx.read_meta_data
+    :return: dictionary with keys 'row', 'col', 'ind', 'shank', 'adc', 'x', 'y', 'sample_shift'
+    """
+    cm = _map_channels_from_meta(meta_data)
+    major_version = _get_neuropixel_major_version_from_meta(meta_data)
+    th = cm.copy()
+    if major_version == 1:
+        # the spike sorting channel maps have a flipped version of the channel map
+        th['col'] = - cm['col'] * 2 + 2 + np.mod(cm['row'], 2)
+    th.update(neuropixel.rc2xy(th['row'], th['col'], version=major_version))
+    th['sample_shift'], th['adc'] = neuropixel.adc_shifts(version=major_version)
+    th['ind'] = np.arange(cm['col'].size)
+    return th
 
 
 def _map_channels_from_meta(meta_data):
@@ -552,6 +587,18 @@ def get_neuropixel_version_from_files(ephys_files):
         return '3B'
     else:
         return '3A'
+
+
+def get_probes_from_folder(session_path):
+    # should glob the ephys files and get out the labels
+    # This assumes the meta files exist on the server (this is the case for now but should it be?)
+    ephys_files = glob_ephys_files(session_path, ext='meta')
+    probes = []
+    for files in ephys_files:
+        if files['label']:
+            probes.append(files['label'])
+
+    return probes
 
 
 def glob_ephys_files(session_path, suffix='.meta', ext='bin', recursive=True, bin_exists=True):
@@ -716,7 +763,9 @@ def download_raw_partial(url_cbin, url_ch, first_chunk=0, last_chunk=0, one=None
     webclient = one.alyx
     cache_dir = cache_dir or webclient.cache_dir
     relpath = Path(url_cbin.replace(webclient._par.HTTP_DATA_SERVER, '.')).parents[0]
-    target_dir = Path(cache_dir, relpath)
+    # write the temp file into a subdirectory
+    tdir_chunk = f"chunk_{str(first_chunk).zfill(6)}_to_{str(last_chunk).zfill(6)}"
+    target_dir = Path(cache_dir, relpath, tdir_chunk)
     Path(target_dir).mkdir(parents=True, exist_ok=True)
 
     # First, download the .ch file if necessary
@@ -747,7 +796,7 @@ def download_raw_partial(url_cbin, url_ch, first_chunk=0, last_chunk=0, one=None
         with open(ch_file_stream, 'r') as f:
             cmeta_stream = json.load(f)
         if (cmeta_stream.get('chopped_first_sample', None) == i0 and
-                cmeta_stream.get('chopped_total_samples', None) == ns_stream):
+                cmeta_stream.get('chopped_total_samples', None) == total_samples):
             return Reader(ch_file_stream.with_suffix('.cbin'))
     else:
         shutil.copy(ch_file, ch_file_stream)