+++ release_notes.md
@@ -1,3 +1,50 @@
+## Release Notes 2.5
+### Release Notes 2.5.1 2021-11-25
+- SpikeSorting task overwrites old tar file on rerun
+### Release Notes 2.5.0 2021-11-24
+- Snapshot class to register images as notes in Alyx
+- Ephys pipeline: RawEphysQC: outputs channels labels as part of pipeline
+- Ephys pipeline: spike sorting
+    - dsp.voltage: destripe detects bad channels and interpolate them
+    - synchronisation 3B: gives a 10% leeway to throw exception
+    - spikes sorting task outputs the RMS plot
+
+## Release Notes 2.4
+### Release Notes 2.4.0
+- Setting tasks to Waiting if they encountered lock (status -2)
+- Setting tasks to Incomplete if they return status -3
+- Completed tasks set held dependent tasks to waiting
+- Adding PostDLC task to compute pupil diameters, licks and DLC QC
+
+## Release Notes 2.3
+### Release Notes 2.3.2 2021-11-08
+- Trial wheel extraction: use alternative sync method when first on fails
+- bugfix: timer for tasks was returning None
+
+### Release Notes 2.3.1 2021-11-5
+- Add setup method in tasks.py into try except to catch globus download errors
+
+### Release Notes 2.3.0 2021-11-4
+- Add input and output signatures to all ephys tasks
+- Add datahandler to task to download and upload data based on location where task is run
+- Spike sorting and EphysVideoSyncQc download data on local servers if not available
+- brainbox.io.one load_spike_sorting_fast: bugfix returns acronyms
+- creates sequence files for spikesorting
+- GPU tasks have a lock - local data handler doesn't instantiate one
+
+
+## Release Notes 2.2
+### Release Notes 2.2.1 2021-11-02
+- allows more than 2 probes in ephys computer probe creation
+
+### Release Notes 2.2.0 2021-10-29
+- brainbox.io.one load_spike_sorting fast: merge channel info in clusters
+- brainbox.io.one generic function to interpolate channels after alignment
+- dsp: spike detection by voltage thresholding and cadzow filtering docs
+- ibllib.io.spikeglx: get geometry from metadata
+- RawEphysQC outputs basic detection spike rates
+- ibllib.atlas.regions: add new mapping cosmos and revise Beryl
+
 ## Release Notes 2.1
 ### Release Notes 2.1.0 2021-10-05
 - destriping as pykilosort internal pre-processing